Added the ability to download a binary response and to test looping workflows
Added the ability to save the response if the content type is set to
FILE, this was mainly added to support downloading binary data files,
but it will work for any response (ignoring text encoding issues).  This
required the updated RestClient with the RawData support.

Added a new child of the RestFixture called the LoopUntilRestFixture
that allows the same REST call to be repeated until a certain state is
reached.  This is designed to enable testing of REST services that take
a significant amount of time to process that return a current status of
processing until complete.